Add Timetable component
