+++ frontend/package.json
@@ -21,6 +21,7 @@
     "react-hook-form": "^7.27.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "type-fest": "^2.12.1",
     "typescript": "^4.5.4",
     "web-vitals": "^2.1.3"
   },
